+++ src/settingsFields.component.jsx
@@ -12,6 +12,7 @@ import PropTypes from 'prop-types'
 import React from 'react'
 import configOptionStore from './configOptionStore.js'
 import Checkbox from './form-fields/check-box.jsx'
+import ColorPicker from './form-fields/color-picker/colorPicker.jsx'
 import SelectField from './form-fields/drop-down.jsx'
 import FileUpload from './form-fields/file-upload.jsx'
 import TextField from './form-fields/text-field.jsx'
@@ -99,12 +100,17 @@ function addConditionallyHiddenStyles(mapping) {
     return settingsValue === currentValue ? { display: 'none' } : {}
 }
 
-function getMenuItems(mapping) {
+function getMenuItems(mapping, apiVersion) {
     const sourceMenuItems =
         (configOptionStore.state && configOptionStore.state[mapping.source]) ||
         []
 
-    const optionsMenuItems = Object.entries(mapping.options || []).map(
+    const options =
+        typeof mapping.options === 'function'
+            ? mapping.options(apiVersion)
+            : mapping.options
+
+    const optionsMenuItems = Object.entries(options || []).map(
         ([id, displayName]) => ({
             id,
             displayName,
@@ -123,7 +129,8 @@ class SettingsFields extends React.Component {
     componentDidMount() {
         this.subscriptions = []
         this.subscriptions.push(
-            settingsStore.subscribe(() => this.forceUpdate())
+            settingsStore.subscribe(() => this.forceUpdate()),
+            configOptionStore.subscribe(() => this.forceUpdate())
         )
     }
 
@@ -140,7 +147,7 @@ class SettingsFields extends React.Component {
         }
     }
 
-    fieldForMapping({ mapping, fieldBase, key, d2 }) {
+    fieldForMapping({ mapping, fieldBase, key, d2, apiVersion }) {
         switch (mapping.type) {
             case 'textfield':
             case undefined:
@@ -160,15 +167,23 @@ class SettingsFields extends React.Component {
                     }),
                 })
 
-            case 'dropdown':
+            case 'dropdown': {
                 if (mapping.includeEmpty && fieldBase.value === '') {
                     fieldBase.value = 'null'
                 }
 
+                const helpText =
+                    typeof mapping.helpText === 'function'
+                        ? mapping.helpText(fieldBase.value, {
+                              configOptions: configOptionStore.getState(),
+                              settings: settingsStore.state,
+                          })
+                        : mapping.helpText
+
                 return Object.assign({}, fieldBase, {
                     component: SelectField,
                     props: Object.assign({}, fieldBase.props, {
-                        menuItems: getMenuItems(mapping),
+                        menuItems: getMenuItems(mapping, apiVersion),
                         includeEmpty: !!mapping.includeEmpty,
                         emptyLabel:
                             (mapping.includeEmpty && mapping.emptyLabel) ||
@@ -177,8 +192,11 @@ class SettingsFields extends React.Component {
                         warning:
                             (mapping.showWarning && mapping.warning) ||
                             undefined,
+                        helpText,
                     }),
                 })
+            }
+
             case 'emailCheckbox': {
                 const emailConfigured = isEmailConfigured(d2)
                 const explanatoryText =
@@ -295,6 +313,23 @@ class SettingsFields extends React.Component {
                     ...fieldBase,
                     component: PeriodTypes,
                 }
+            case 'colorPicker':
+                return {
+                    ...fieldBase,
+                    component: ColorPicker,
+                    props: {
+                        label: fieldBase.props.floatingLabelText,
+                        color: fieldBase.value,
+                        onColorPick: ({ color }) => {
+                            settingsActions.saveKey(key, color)
+                            settingsActions.saveKey(
+                                'keyCustomColorMobile',
+                                color
+                            )
+                        },
+                    },
+                }
+
             default:
                 console.warn(
                     `Unknown control type "${mapping.type}" encountered for field "${key}"`
@@ -351,7 +386,13 @@ class SettingsFields extends React.Component {
                     validators,
                 }
 
-                return this.fieldForMapping({ mapping, fieldBase, key, d2 })
+                return this.fieldForMapping({
+                    mapping,
+                    fieldBase,
+                    key,
+                    d2,
+                    apiVersion: this.props.apiVersion,
+                })
             })
             .filter((f) => f && !!f.name)
             .map((field) => {
@@ -407,6 +448,7 @@ class SettingsFields extends React.Component {
 }
 
 SettingsFields.propTypes = {
+    apiVersion: PropTypes.number.isRequired,
     category: PropTypes.string.isRequired,
     currentSettings: PropTypes.arrayOf(PropTypes.string).isRequired,
 }